feat: update version to 1.3.2 and adjust cooldown time calculations

+++ gameConnection/sceneUser2Msg.go
@@ -25,7 +25,7 @@ func (g *GameConnection) HandleSceneUser2ProtoCmd(cmdParamId int32, rawData []by
 	case Cmd.User2Param_value["USER2PARAM_SERVERTIME"]:
 		param = &Cmd.ServerTime{}
 		err = utils.ParseCmd(rawData, param)
-		fixedSkillCDSubtract = param.(*Cmd.ServerTime).GetTime() - uint64(time.Now().UnixMilli())
+		fixedSkillCDSubtract = int64(param.(*Cmd.ServerTime).GetTime() - uint64(time.Now().UnixMilli()))
 
 	case Cmd.User2Param_value["USER2PARAM_GAMETIME"]:
 		param = &Cmd.GameTimeCmd{}
@@ -76,7 +76,7 @@ func (g *GameConnection) HandleSceneUser2ProtoCmd(cmdParamId int32, rawData []by
 						continue
 					}
 					now := time.Now().UnixMilli()
-					calculatedCd := int64(cd.GetTime() - fixedSkillCDSubtract)
+					calculatedCd := int64(cd.GetTime()) - fixedSkillCDSubtract
 					if calculatedCd < now {
 						continue
 					}
